dynamically store contract addresses in mysql
When `truffle migrate` runs, it dumps build manifest in JSON format, containing
the addresses for the deployed contract. A script goes through those, finds the
right address for each contract (local RPC testnet is hardcoded to network ID
999999).
It makes PUT to the backend, which stores them in the DB. The frontend can then
request them via an API endpoint.